pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
name = "mailingester"
version = "0.1.0"
description = "Extract information from emails"
authors = [
    {name = "IDEMS International", email = "communications@idems.international"},
]
readme = "README.md"
requires-python = ">=3.11"
keywords = []
license = {text = "GPL-3.0-or-later"}
classifiers = []
dependencies = [
    "aiosmtpd",
    "google-cloud-storage",
    "pydantic-settings",
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "isort",
    "pytest",
]
build = [
    "build",
]

[project.urls]
Homepage = "https://github.com/IDEMSInternational/mailingester"
Repository = "https://github.com/IDEMSInternational/mailingester"
